I am now formatting ICD-10-CM codes in the 'dot' format when they go out to GPT for re-ranking. This is an experiment. When you ask GPT normal questions about ICD-10 medical codes, it tends to reference codes in the dot format. This is generally thought of as the more human-readable format. So, we'll see if using this format in the GPT prompt makes any difference in the suggestions it makes.
I am also formatting the codes back to the CMS standard (no dot) format in the re-ranking response payload.
Once I started sending the codes to GPT in the dot format (see above), it began using the dot format in the re-ranked responses. So we need to guard against that by formatting them on the way in as well (to CMS format - no dot).